pyproject.toml
[project]
name = "secureorder-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.14.5",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "starlette>=0.52.1",
    "uvicorn>=0.40.0",
]
